Allows multiple .sol files for --sol option. Moved chifra abis --find to ./chifra/cmd/abis folder. Remove go-apps version.
