+++ src/go-apps/CMakeLists.txt
@@ -15,4 +15,3 @@ add_subdirectory(blaze)
 add_subdirectory(flame)
 add_subdirectory(acctScrape2)
 # add_subdirectory(unchained)
-add_subdirectory(findSig)
