--- src/go-apps/findSig/go.mod
@@ -1,10 +1,0 @@
-module github.com/TrueBlocks/trueblocks-core/src/go-apps/findSig
-
-go 1.15
-
-require (
-	github.com/ethereum/go-ethereum v1.10.9
-	github.com/mitchellh/go-homedir v1.1.0
-	github.com/panjf2000/ants/v2 v2.4.1
-	github.com/spf13/cobra v0.0.5
-)
