+++ test/gold/tools/grabABI/grabABI_fromSolidity.txt
@@ -10,8 +10,8 @@ Arguments:
 
 Flags:
   -k, --known          load common 'known' ABIs from cache
-  -s, --sol string     file name of .sol file from which to create a new known abi (without .sol)
-  -f, --find strings   try to search for a function declaration given a four byte code
+  -s, --sol string     extract the abi definition from the provided .sol file(s)
+  -f, --find strings   search for function or event declarations given a four- or 32-byte code(s)
   -o, --source         show the source of the ABI information (hidden)
   -c, --classes        generate classDefinitions folder and class definitions (hidden)
